Refactor FDBDatabase to FDBDatabaseImpl so that it's behind an interface (use interface everywhere if possible). Refactor FDBDatabaseFactory to FDBDatabaseFactoryImpl so that it's behind an interface (use interface everywhere if possible).
