+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBRecordContext.java
@@ -78,7 +78,7 @@ import java.util.stream.Collectors;
  * </p>
  *
  * <pre><code>
- * final FDBDatabase fdb = FDBDatabaseFactory.instance().getDatabase();
+ * final FDBDatabase fdb = FDBDatabaseFactoryImpl.instance().getDatabase();
  * try (FDBRecordContext ctx = fdb.openContext()) {
  *     ...
  * }
