+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/FDBTypedRecordStoreTest.java
@@ -67,7 +67,7 @@ public class FDBTypedRecordStoreTest extends FDBTestBase {
 
     @BeforeEach
     public void setup() {
-        fdb = FDBDatabaseFactory.instance().getDatabase();
+        fdb = FDBDatabaseFactoryImpl.instance().getDatabase();
         try (FDBRecordContext context = fdb.openContext()) {
             FDBRecordStore.deleteStore(context, TestKeySpace.getKeyspacePath("record-test", "unit", "typedtest"));
             context.commit();
